tagsiyacht/requirements.txt
Django==1.8.1
djangorestframework==3.1.1
Pillow==2.8.1
psycopg2==2.6
